Refactor JS for index page
Utilized initializePage function
Import right index.js in both openconext/base
