+++ theme/base/javascripts/application.js
@@ -1,5 +1,5 @@
-import {initializeIndex} from ".";
-import {initializeWayf} from ".";
+import {initializeIndex} from "./index";
+import {initializeWayf} from "./wayf";
 import {initializeConsent} from './consent';
 
 initializeIndex();
