+++ theme/base/javascripts/index.js
@@ -1,10 +1,6 @@
 import {replaceMetadataCertificateLinkTexts} from "./modules/EngineBlockMainPage";
+import {initializePage} from './page';
 
 export function initializeIndex() {
-    document.body.className = document.body.className.replace('no-js', '');
-
-    if (document.getElementById('engine-main-page') !== null) {
-        replaceMetadataCertificateLinkTexts();
-        return;
-    }
+  initializePage('#engine-main-page', replaceMetadataCertificateLinkTexts);
 }
