+++ theme/openconext/javascripts/application.js
@@ -1,4 +1,4 @@
-import {initializeIndex} from "../../base/javascripts";
+import {initializeIndex} from "./index";
 import {initializeWayf} from "../../base/javascripts/wayf";
 import {initializeConsent} from './consent';
